--- deploy/kava_mainnet/111_deploy_Kava2Pool.ts
@@ -1,59 +1,0 @@
-import { HardhatRuntimeEnvironment } from "hardhat/types"
-import { DeployFunction } from "hardhat-deploy/types"
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments, getNamedAccounts, getChainId } = hre
-  const { execute, get, getOrNull, log, read, save } = deployments
-  const { deployer } = await getNamedAccounts()
-
-  // Manually check if the pool is already deployed
-  const saddleKava2pool = await getOrNull("SaddleKava2pool")
-  if (saddleKava2pool) {
-    log(`reusing "Kava2poolTokens" at ${saddleKava2pool.address}`)
-  } else {
-    // Constructor arguments
-    const TOKEN_ADDRESSES = [
-      (await get("USDC")).address,
-      (await get("USDT")).address,
-    ]
-    const TOKEN_DECIMALS = [6, 6]
-    const LP_TOKEN_NAME = "Saddle 2pool"
-    const LP_TOKEN_SYMBOL = "saddleKavaUSD"
-    const INITIAL_A = 400
-    const SWAP_FEE = 4e6 // 4bps
-    const ADMIN_FEE = 5e9
-
-    await execute(
-      "SwapFlashLoan",
-      { from: deployer, log: true },
-      "initialize",
-      TOKEN_ADDRESSES,
-      TOKEN_DECIMALS,
-      LP_TOKEN_NAME,
-      LP_TOKEN_SYMBOL,
-      INITIAL_A,
-      SWAP_FEE,
-      ADMIN_FEE,
-      (
-        await get("LPToken")
-      ).address,
-    )
-
-    await save("SaddleKava2pool", {
-      abi: (await get("SwapFlashLoan")).abi,
-      address: (await get("SwapFlashLoan")).address,
-    })
-
-    const lpTokenAddress = (await read("SaddleKava2pool", "swapStorage"))
-      .lpToken
-    log(`Saddle Kava USD Pool LP Token at ${lpTokenAddress}`)
-
-    await save("SaddleKava2poolLPToken", {
-      abi: (await get("LPToken")).abi, // LPToken ABI
-      address: lpTokenAddress,
-    })
-  }
-}
-export default func
-func.tags = ["SaddleKava2pool"]
-func.dependencies = ["SwapUtils", "SwapFlashLoan", "Kava2poolTokens"]
